+++ test/exop_test.exs
@@ -1,3 +1,0 @@
-defmodule ExopTest do
-  use ExUnit.Case
-end